Reorganize Tests project
