+++ SpecFlow.DependencyInjection.Tests/Steps/ContextInjectionScopeSteps.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using TechTalk.SpecFlow;
 using Xunit;
 
-namespace SolidToken.SpecFlow.DependencyInjection.Tests
+namespace SolidToken.SpecFlow.DependencyInjection.Tests.Steps
 {
     public class TestContext
     {
