+++ SpecFlow.DependencyInjection.Tests/Steps/ISpecFlowOutputHelperSteps.cs
@@ -7,7 +7,7 @@ using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Infrastructure;
 using Xunit;
 
-namespace SolidToken.SpecFlow.DependencyInjection.Tests
+namespace SolidToken.SpecFlow.DependencyInjection.Tests.Steps
 {
     [Binding]
     public class ISpecFlowOutputHelperSteps
